+++ src/pacx/clients/__init__.py
@@ -1,9 +1,11 @@
 from .connectors import ConnectorsClient as ConnectorsClient
 from .dataverse import DataverseClient as DataverseClient
+from .policy import DataLossPreventionClient as DataLossPreventionClient
 from .power_platform import PowerPlatformClient as PowerPlatformClient
 
 __all__ = [
     "ConnectorsClient",
     "DataverseClient",
+    "DataLossPreventionClient",
     "PowerPlatformClient",
 ]